feat(cli): retrieval source-mode, advisory-lock management, orient --fast, and init exit codes
A coordinated agent-UX and operational-safety wave across the context/CLI
core. These changes are committed together because they are mutually
compile-coupled through the shared `ContextPackOptions` struct, `cli/mod.rs`,
and the provenance/degradation model types.

Retrieval source mode
- Add `--source-mode lexical_only|semantic_only|hybrid` and
  `--strict-source-mode` to `ee pack`, `ee pack build`, and `ee context`.
  `ContextPackOptions` gains `source_mode`/`strict_source_mode`, threaded
  through `run_context_pack` into the search request (replacing the
  previously hardcoded `Hybrid`/non-strict pair). Strict mode fails instead
  of silently falling back when the requested source is unavailable.
- The performance/query-plan JSON now reports `sourceModeRequested`,
  `sourceModeApplied`, `strictSourceMode`, and `fallbackApplied`.
- The L2 pack cache keys on source mode and strict flag, refuses to serve or
  store entries produced by a source-mode fallback, and records the source
  mode alongside the cached payload. All bench/test `ContextPackOptions`
  constructors set the new fields.

Advisory-lock management (`ee diag advisory-lock`)
- Add `--list`, `--release`, and `--force`. `--list` reports active locks
  with holder liveness; `--release` reclaims a dead same-host holder;
  `--force` releases even when liveness is unknown/alive. Backed by new
  db surfaces (`advisory_lock_holder_liveness`, `list_active_advisory_locks`,
  `force_release_advisory_lock`, `release_reclaimable_advisory_lock`,
  `AdvisoryLockReleaseOutcome`). Repair hints across `remember`, `curate`,
  the `advisory_lock_timeout` fixture, and `degraded_codes.md` now point at
  the runnable `--resource-type ... --release` form.

orient --fast
- `ee orient` gains `--fast` (alias `--quick`): a bounded session-start
  bundle that skips full doctor and the embedded pack, caps the swarm-brief
  timeout, and emits `orient_doctor_skipped`/`orient_pack_skipped` info
  degradations with explicit follow-up commands. Orient's suggested pack
  command now uses `--read-only --source-mode lexical_only`.

init exit codes & symlink resolution
- `ee init` returns a non-zero exit (storage vs configuration) on failed
  actions instead of always exiting success; `InitReport` gains
  `action_errors` with per-action reasons (surfaced in human + JSON output).
- Harmless ancestor symlinks (e.g. macOS `/tmp` -> `/private/tmp`) are
  canonicalized before the no-follow guard runs, so scratch workspaces no
  longer require `realpath` busywork; a final-component symlink is still
  refused unless `--allow-symlink`.

Provenance & misc
- Add the `manual://` / external provenance scheme (`ProvenanceUri::External`),
  handled in evidence-freshness assessment as an accepted-but-not-file-checkable
  source; ADR 0043 and the derived-candidate contract test track the renamed
  `provenance_uri_accepts_registered_schemes` pin.
- preflight guard: treat cargo `-Z` as a value-taking global option so
  `cargo -Z unstable-options ...` parses correctly.
- `--fields` help notes it may appear before or after subcommands; workspace
  and advisory-lock JSON responses carry a top-level `degraded: []`.
- Rename the `context_stream` bench target to `context_stream_bench` with an
  explicit path.

+++ tests/e2e_lab_swarm_workload_generator.rs
@@ -135,7 +135,7 @@ fn lab_generate_workload_emits_all_profiles_as_redaction_safe_json() -> TestResu
             format!("generator evidence profile mismatch for {profile}: {value}"),
         )?;
         ensure(
-            value["generatorEvidence"]["commandCount"].as_u64() == Some(command_count),
+            value["generatorEvidence"]["commandCount"].as_u64() == Some(command_count as u64),
             format!("generator evidence command count mismatch for {profile}: {value}"),
         )?;
         ensure(
